Add files via upload
This is my own worked solution for Project Euler for Question 11, https://projecteuler.net/problem=11

It's not easy to extract the information from STRING to Numbers, so I created a function to convert strings to vectorized input. And I compiled my own 20 X 20 Matrix using my Extract function customized for the String Information for that question, and later compile the 20 X 20 Matrix, by using ; to vertical concatenate the vectors. I spent about 4 hours to develop the algorithm to find the largest 4 number sum.